+++ internal/params/struct.go
@@ -215,10 +215,28 @@ func (s *structValue) JSON(v string) *structure {
 	return s.parent
 }
 
+func (s *structValue) JSONFromBytes(v []byte) *structure {
+	s.parent.values = append(s.parent.values, value.StructValueField{
+		Name: s.name,
+		V:    value.JSONValueFromBytes(v),
+	})
+
+	return s.parent
+}
+
 func (s *structValue) JSONDocument(v string) *structure {
 	s.parent.values = append(s.parent.values, value.StructValueField{
 		Name: s.name,
 		V:    value.JSONDocumentValue(v),
+	})
+
+	return s.parent
+}
+
+func (s *structValue) JSONDocumentFromBytes(v []byte) *structure {
+	s.parent.values = append(s.parent.values, value.StructValueField{
+		Name: s.name,
+		V:    value.JSONDocumentValueFromBytes(v),
 	})
 
 	return s.parent