+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* Added optimized helper methods for efficient `string` <-> `[]byte` conversions in `JSON` and `JSONDocument` types.
+* Added optimized helper methods for efficient `[]byte` → `string` conversions in `JSON` and `JSONDocument` types.
 
 ## v3.125.2
 * Added public named errors `query.{ErrNoRows,ErrMoreThanOneRow,ErrMoreThanOneResultSet,ErrNoResultSets}` in the `query` package